fix: force render
